Revamp of blocks + statements.
Redo the implementation of Bstatement/Bblock to accommodate the way in
which these classes are used and manipulated in the front end.  In the
gcc backend the underlying data structures for
Bblock/Bstatement/Bexpression are all tree fragments, meaning that it
can (and does) treat blocks and statements as interchangeable, which
doesn't play well with equivalents in LLVM (like llvm::BasicBlock).

In the new scheme, we capture the block structure of the program as
the frontend creates blocks and statements, then go through and
materializat control flow as a second pass (creating blocks, assigning
instructions to blocks, adding conditional jumps, etc). This requires
more scaffolding, but allows things to work without a lot of changes in
the frontend code.

No support yet for statements that involve control flow, that will come
as a second step.