Distinguish between lvalue + rvalue contexts for var expressions.
Change gofrontend to pass in lvalue/rvalue context for var_expressions,
so that we know whether we want a raw var address or a var load. For
example, in the statement

   x, y.arr[z].f = q, r

the variables x and y appear in an lvalue context whereas z, q, and r do
not. Having this info passed in from the front end removes the need to
discover this info in the back end later on.

Warning: this change requires a corresponding change in gofrontend
that has not yet been submitted.