Work around issue in Llvm_backend::address_expression.
The helper String_expression::do_get_backend() in gofrontend
seems to want to take the address of a string constant-- this causes
issues in the LLVM backend since string values are already
effectively pointers. For now just treat the "&" operator as a no-op
if we encounter it in Llvm_backend::address_expression.