Add bitcasts in when creating composite constants.
Revise the helper routine Llvm_backend::makeConstCompositeExpr to
check for cases where the types of composite element values do not
match the expected type of the composite element or field.  I hate to
do this (since by adding conversions we could be papering over actual
bugs), but it seems to be hard to avoid due to the use of placeholder
types in the front end.