Assorted changes.
Completely rework the support for circular pointer types, since it
was clear that it wasn't set up properly to handle cycles of size
more than one, e.g.

  type p *q
  type q *p

Add support for condition_expression and compound_expression. Improve
trace output for type construction. Next step is going to be a revamp
of placeholder types, since the work on circular types made it clear
that there are some issues with placeholders as well.