Overhaul of placeholder type management.
Reimplement the way placeholder types are handled; this new scheme is a
good deal more robust. I've also done away with a bunch of side
tables for rtype management at the backend level and instead
moved to a Btype class hierarchy (with the equivalent info stored in
Btype derived classes).