Fix string commoning.
The existing mechanism for commoning of constants was not working for
string constants. LLVM string constants have array type (ex: "[5 x
i8]") whereas it is more convenient for string Bexpressions to have
pointer type ("i8*").  To achieve this the string constant builder
creates a module-scope var initialized with the the constant, then
creates a bitcast of the var's value to the correct type. The problem
here is that by creating a new var each time, the caching is defeated
(each module var has a new value). I fixed this by adding a new
separate map for strings, keyed by llvm::Value of the string constant.